requirements.txt
annotated-types==0.6.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
docstring_parser==0.16
filelock==3.13.4
fsspec==2024.3.1
gcloud==0.18.3
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-aiplatform==1.48.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.62.2
httplib2==0.22.0
huggingface-hub==0.22.2
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
oauth2client==4.1.3
packaging==24.0
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
pypdfium2==4.29.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rsa==4.9
safetensors==0.4.3
shapely==2.0.4
six==1.16.0
sympy==1.12
tokenizers==0.19.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.40.0
typing_extensions==4.11.0
urllib3==2.2.1
gitpython==3.1.43
